+++ src/aioredis_cluster/factory.py
@@ -3,7 +3,6 @@ import warnings
 from typing import Any, Dict, List, Optional, Sequence
 
 import aioredis
-from aioredis.abc import AbcPool
 from aioredis.util import parse_url
 
 from aioredis_cluster.abc import AbcCluster
@@ -17,20 +16,6 @@ __all__ = [
     "create_cluster",
     "create_redis_cluster",
 ]
-
-
-class PoolFactory:
-    def __init__(self, default_pool_opts: Optional[Dict[str, Any]] = None):
-        if default_pool_opts is None:
-            default_pool_opts = {}
-        self._default_pool_opts: Dict[str, Any] = default_pool_opts
-
-    async def create_pool(self, addr: AioredisAddress, opts: Dict[str, Any] = None) -> AbcPool:
-        if opts is None:
-            opts = {}
-        pool = await create_pool(addr, **{**self._default_pool_opts, **opts})
-
-        return pool
 
 
 async def create_cluster(
